connected backend for profiles and added bug fixes

+++ src/components/companion/blueprintProfiles/additionalLinks.tsx
@@ -4,7 +4,7 @@ import ChainLinkIcon from "../../../../public/assets/icons/chain_link.svg";
 import Image from 'next/image'
 
 interface UserProfile {
-    additionalLink?: string;
+    additionalLink?: string | string[];
 }
 
 interface AdditionalLinksProps {
@@ -12,26 +12,43 @@ interface AdditionalLinksProps {
 }
 
 const AdditionalLinks: FC<AdditionalLinksProps> = ({ userData }) => {
+
+    const renderLinks = () => {
+        if (!userData.additionalLink) return null;
+
+        const links = Array.isArray(userData.additionalLink) ? userData.additionalLink : [userData.additionalLink];
+
+        return (
+            <ul className="list-none p-0 m-0">
+                {links.map((link, index) => (
+                    <li key={index} className="mb-2">
+                        <a
+                            href={link}
+                            className="text-xs sm:text-sm font-satoshi underline break-all hover:text-purple-400 transition-colors max-w-none sm:max-w-none truncate flex"
+                        >
+                            <Image src={ChainLinkIcon}
+                                alt="Linkedin Icon"
+                                width={16}
+                                height={16}
+                                className='mr-2'
+                            />
+                            {link}
+                        </a>
+                    </li>
+                ))}
+            </ul>
+        );
+    };
+
     return (
         <AnimatedBorder className="w-full mb-3 sm:mb-4">
             <div className="rounded-lg p-3 sm:p-4">
-                <span className="text-xs sm:text-sm text-light-grey font-redhat mb-1 sm:mb-2">ADDITIONAL LINK</span>
+                <span className="text-xs sm:text-sm text-light-grey font-redhat mb-1 sm:mb-2">ADDITIONAL LINKS</span>
                 <div className="flex justify-between items-center">
-                    <a
-                        href={userData.additionalLink}
-                        className="text-xs sm:text-sm font-satoshi underline break-all hover:text-purple-400 transition-colors max-w-none sm:max-w-none truncate flex"
-                    >
-                        <Image src={ChainLinkIcon}
-                            alt="Linkedin Icon"
-                            width={16}
-                            height={16}
-                            className='mr-2'
-                        />
-                        {userData.additionalLink}
-                    </a>
+                    {renderLinks()}
                 </div>
             </div>
-        </ AnimatedBorder>
+        </AnimatedBorder>
     );
 };
